Code Refactor | Update Method and Class name and their responsibility to be more meaningful
