+++ readme.md
@@ -1,11 +1,11 @@
 # Problem:: Create a TIC-TOE game engine
 
 ## Objective::
-    This implementation is initial set of working copy. There is still improvement opportunity related
-    design and code readibility prespective.
-    In subsequent promote it be more robust, which will be based on design priniciple
+
+    This Version will adhere to SOLID priniciple
 
 ## Testing Steps::
+
 - Provide row and col number
 - Check current status of the game
 - Start over again till gave is over
