+++ src/main/java/Main.java
@@ -1,4 +1,6 @@
+import org.demo.api.AIEngine;
 import org.demo.api.GameEngine;
+import org.demo.api.RuleEngine;
 import org.demo.boards.TicTacToeBoard;
 import org.demo.game.Board;
 import org.demo.game.Cell;
@@ -15,29 +17,31 @@ public class Main
         {
             Scanner scanner = new Scanner(System.in);
             GameEngine gameEngine = new GameEngine();
+            RuleEngine ruleEngine = new RuleEngine();
+            AIEngine aiEngine = new AIEngine();
             Board board = gameEngine.start("TicTacToe");
 
-            while (!gameEngine.isComplete(board).isOver())
+            while (!ruleEngine.getState(board).isOver())
             {
                 System.out.println("Current Status::\n" + (TicTacToeBoard) board);
                 System.out.println("Make Your Move");
 
                 Player computer = new Player("O");
-                Player opponent = new Player("X");
+                Player human = new Player("X");
 
                 int row = scanner.nextInt();
                 int col = scanner.nextInt();
 
-                Move oppMove = new Move(new Cell(row, col));
-                gameEngine.move(board, opponent, oppMove);
+                Move oppMove = new Move(new Cell(row, col), human);
+                gameEngine.move(board, oppMove);
 
-                if (!gameEngine.isComplete(board).isOver())
+                if (!ruleEngine.getState(board).isOver())
                 {
-                    Move computerMove = gameEngine.suggestMove(computer, board);
-                    gameEngine.move(board, computer, computerMove);
+                    Move computerMove = aiEngine.suggestMove(computer, board);
+                    gameEngine.move(board, computerMove);
                 }
             }
-            System.out.println("GameResult: " + gameEngine.isComplete(board));
+            System.out.println("GameResult: " + ruleEngine.getState(board));
             System.out.println("FinalResult::\n" + (TicTacToeBoard) board);
             System.out.println("Gave is Over!");
         }
