+++ src/main/java/org/demo/game/Board.java
@@ -1,6 +1,6 @@
 package org.demo.game;
 
-public class Board
+public abstract class Board
 {
-
+    public abstract void move(Move move);
 }
