+++ src/main/java/org/demo/game/GameState.java
@@ -1,11 +1,11 @@
 package org.demo.game;
 
-public class GameResult
+public class GameState
 {
     boolean isOver;
     String winner;
 
-    public GameResult(boolean isOver, String winner)
+    public GameState(boolean isOver, String winner)
     {
         this.isOver = isOver;
         this.winner = winner;
