+++ src/main/java/org/demo/game/Move.java
@@ -4,15 +4,23 @@ public class Move
 {
     private Cell cell;
 
+    private Player player;
 
-    public Move(Cell cell)
+
+    public Move(Cell cell, Player player)
     {
         this.cell = cell;
+        this.player = player;
     }
 
     public Cell getCell()
     {
         return cell;
+    }
+
+    public Player getPlayer()
+    {
+        return player;
     }
 }
